Moving image processing from dataset to image.py
